requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
Pillow>=10.2.0
docx2pdf>=0.1.8
comtypes>=1.2.0
Flask-Cors>=4.0.0
